Abstracted DateBase class out of DateOfBirth class
